Refactor SerializationGroups config
- move into a parent `controllers` config.

This allows us to specify which controllers we want to use.
Trying to create all controllers in an application is dangerous,
as there may be some in 3rd party modules which won't start.

Specifically, DoctrineORM module has some debug controllers
which will fail on instantiation if debug is not enabled.